+++ crates/snow-native/Cargo.toml
@@ -10,3 +10,5 @@ crate-type = ["cdylib"]
 [dependencies]
 napi = { version = "3", features = ["napi9"] }
 napi-derive = "3"
+encoding_rs = "0.8"
+chardetng = "0.1"
