+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "snow-ai",
-	"version": "0.8.20",
+	"version": "0.8.21",
 	"description": "Agentic coding in your terminal",
 	"license": "MIT",
 	"bin": {
