+++ source/mcp/utils/filesystem/read-tools.utils.ts
@@ -11,6 +11,7 @@ import {
 	readFileLinesStreaming,
 	readFileWithEncoding,
 } from './encoding.utils.js';
+import {readFileWithNative} from './native-edit.utils.js';
 import {readOfficeDocument} from './office-parser.utils.js';
 import {formatLineWithHash} from './hashline.utils.js';
 
@@ -22,14 +23,11 @@ type GetFileContentContext = {
 	isSSHPath: (filePath: string) => boolean;
 	readRemoteFile: (sshUrl: string) => Promise<string>;
 	isImageFile: (filePath: string) => boolean;
-	readImageAsBase64: (fullPath: string) => Promise<
-		| {
-				type: 'image';
-				data: string;
-				mimeType: string;
-		  }
-		| null
-	>;
+	readImageAsBase64: (fullPath: string) => Promise<{
+		type: 'image';
+		data: string;
+		mimeType: string;
+	} | null>;
 	isOfficeFile: (filePath: string) => boolean;
 	getNotebookEntries: (filePath: string) => string;
 	extractRelevantSymbols: (
@@ -162,7 +160,10 @@ export async function executeGetFileContentCore(
 					lines = streamed.lines;
 					totalLines = streamed.totalLines;
 				} else {
-					content = await readFileWithEncoding(fullPath);
+					content = (await readFileWithNative(fullPath)) ?? '';
+					if (!content) {
+						content = await readFileWithEncoding(fullPath);
+					}
 					lines = content.split('\n');
 					totalLines = lines.length;
 				}
@@ -194,11 +195,17 @@ export async function executeGetFileContentCore(
 					fileSizeBytes > FILE_SIZE_LIMIT
 						? ` [Large file: ${Math.round(fileSizeBytes / 1024 / 1024)}MB]`
 						: '';
-				let fileContent = `${file} (lines ${start}-${end}/${totalLines})${sizeWarning}\n${numberedLines.join('\n')}`;
+				let fileContent = `${file} (lines ${start}-${end}/${totalLines})${sizeWarning}\n${numberedLines.join(
+					'\n',
+				)}`;
 
 				if (content) {
 					try {
-						const symbols = await parseFileSymbols(fullPath, content, ctx.basePath);
+						const symbols = await parseFileSymbols(
+							fullPath,
+							content,
+							ctx.basePath,
+						);
 						const symbolInfo = ctx.extractRelevantSymbols(
 							symbols,
 							start,
@@ -219,10 +226,17 @@ export async function executeGetFileContentCore(
 				}
 
 				multimodalContent.push({type: 'text', text: fileContent});
-				filesData.push({path: file, startLine: start, endLine: end, totalLines});
+				filesData.push({
+					path: file,
+					startLine: start,
+					endLine: end,
+					totalLines,
+				});
 			} catch (error) {
-				const errorMsg = error instanceof Error ? error.message : 'Unknown error';
-				const inputPath = typeof fileItem === 'string' ? fileItem : fileItem.path;
+				const errorMsg =
+					error instanceof Error ? error.message : 'Unknown error';
+				const inputPath =
+					typeof fileItem === 'string' ? fileItem : fileItem.path;
 				let resolvedPathInfo = '';
 				try {
 					const attemptedResolve = ctx.resolvePath(inputPath, lastAbsolutePath);
@@ -344,11 +358,16 @@ export async function executeGetFileContentCore(
 		lines = streamed.lines;
 		totalLines = streamed.totalLines;
 		const start = Math.min(actualStartLine, totalLines);
-		const end = Math.min(totalLines, Math.min(actualEndLine, start + lines.length - 1));
+		const end = Math.min(
+			totalLines,
+			Math.min(actualEndLine, start + lines.length - 1),
+		);
 		const numberedLines = lines.map((line, index) =>
 			formatLineWithHash(start + index, line),
 		);
-		const sizeInfo = `[File: ${Math.round(fileSizeBytes / 1024 / 1024)}MB, ${totalLines} lines total. Showing lines ${start}-${end}. Use startLine/endLine to read other sections.]`;
+		const sizeInfo = `[File: ${Math.round(
+			fileSizeBytes / 1024 / 1024,
+		)}MB, ${totalLines} lines total. Showing lines ${start}-${end}. Use startLine/endLine to read other sections.]`;
 		return {
 			content: `${sizeInfo}\n${numberedLines.join('\n')}`,
 			startLine: start,
@@ -357,7 +376,10 @@ export async function executeGetFileContentCore(
 		};
 	}
 
-	content = await readFileWithEncoding(fullPath);
+	content = (await readFileWithNative(fullPath)) ?? '';
+	if (!content) {
+		content = await readFileWithEncoding(fullPath);
+	}
 	lines = content.split('\n');
 	totalLines = lines.length;
 	const actualStartLine = startLine ?? 1;
@@ -378,7 +400,12 @@ export async function executeGetFileContentCore(
 	let partialContent = numberedLines.join('\n');
 	try {
 		const symbols = await parseFileSymbols(fullPath, content, ctx.basePath);
-		const symbolInfo = ctx.extractRelevantSymbols(symbols, start, end, totalLines);
+		const symbolInfo = ctx.extractRelevantSymbols(
+			symbols,
+			start,
+			end,
+			totalLines,
+		);
 		if (symbolInfo) {
 			partialContent += symbolInfo;
 		}
